Rename 1.19 and 1.20 packages
