+++ pom.xml
@@ -15,8 +15,8 @@
     <modules>
         <module>modules/api</module>
         <module>modules/v1_18</module>
-        <module>modules/v1_19_R3</module>
-        <module>modules/v1_20_R4</module>
+        <module>modules/v1_19</module>
+        <module>modules/v1_20</module>
         <module>modules/datapack</module>
         <module>modules/Movecraft</module>
     </modules>
